abstract permissions
