+++ subgraph/src/index.ts
@@ -1,2 +1,3 @@
 export * from './ids/dao';
 export * from './ids/pluginRepo';
+export * from './ids/permissions';
